+++ test/php/src/demo/Input_Data.php
@@ -115,7 +115,7 @@ class Input_Data {
             case 'int[]':
                 $val = [];
                 foreach ($strValue as $str) {
-                    $v = self::parseValue('int64', $str);
+                    $v = self::parseValue('int', $str);
                     if ($v === null) return null;
                     $val[] = $v;
                 }
